+++ R/getPlayerCompass.R
@@ -23,7 +23,11 @@ getPlayerCompass <- function(player_id=NULL, n_compass_points=c("16","8","4"))
     n_compass_points <- as.character(n_compass_points)
     n_compass_points <- match.arg(n_compass_points)
 
-    rot <- miner::getPlayerRotation(player_id)
+    rot <- NA
+    while(is.na(rot)) rot <- miner::getPlayerRotation(player_id)
+    while(rot < 0) rot <- rot + 360 # force to value >= 0
+    rot <- rot %% 360 # force to value 0-360
+    rot <- rot - 360 # force to value -360 - 0
 
     rotation_to_compass(rot, n_compass_points)
 }